+++ HISTORY.rst
@@ -12,6 +12,7 @@ History
 * Add publish button
 * Fix issues in migrations. Thanks @skirsdeda
 * Fix selecting current menu item according to menu layout
+* Add separated fields for social network descriptions
 
 0.6.3 (2015-12-22)
 ++++++++++++++++++